+++ Puppetfile.tracking
@@ -357,8 +357,8 @@ mod 'simp-selinux',
   :tag => '2.1.2'
 
 mod 'simp-simp',
-  :git => 'https://github.com/simp/pupmod-simp-simp',
-  :tag => '4.4.1'
+  :git => 'https://github.com/jeefberkey/pupmod-simp-simp',
+  :branch => 'SIMP-4857-ipa-enrollment'
 
 mod 'simp-simpcat',
   :git => 'https://github.com/simp/pupmod-simp-simpcat',
