+++ spec/acceptance/suites/ipa/40_ipa_spec.rb
@@ -1,37 +1,38 @@
 require 'spec_helper_integration'
 
-def run_ipa_cmd(host, pass)
-  on(ipa, "echo \"#{pass}\" | kinit admin")
-  result = on(ipa, cmd)
-  on(ipa, 'kdestroy')
+def run_ipa_cmd(host, pass, cmd)
+  on(host, "echo \"#{pass}\" | kinit admin")
+  result = on(host, cmd)
+  on(host, 'kdestroy')
 
   result
 end
 
 describe 'validate the ipa server' do
 
-  ipa_server  = hosts_with_role(hosts, 'ipa_server').first
-  ipa_clients = hosts_with_role(hosts, 'ipa_client')
+  admin_password = '@dm1n=P@ssw0r'
+  ipa_server     = hosts_with_role(hosts, 'ipa_server').first
+  ipa_clients    = hosts_with_role(hosts, 'ipa_client')
 
   context 'server' do
     it 'should have 4 hosts in the inventory' do
-      out   = run_ipa_cmd(ipa_server, 'ipa host-find')
+      out   = run_ipa_cmd(ipa_server, admin_password, 'ipa host-find')
       hosts = out.stdout.split("\n").grep(/Host name/)
 
       expect(hosts.length).to eq 4
     end
 
     it 'should have dns entries for each host' do
-      out     = run_ipa_cmd(ipa_server, 'ipa dnsrecord-find test.case')
+      out     = run_ipa_cmd(ipa_server, admin_password, 'ipa dnsrecord-find test.case')
       records = out.stdout.split("\n").grep(/Record name/).map {|h|h.split(': ').last}
 
       expect(records).to include %w[ puppet ipa el6-client el7-client ]
     end
 
     it 'should add a test user and a posix group' do
-      out = run_ipa_cmd(ipa_server, 'ipa user-add testuser --first=Test --last=User --displayname="Test User" --random')
-      run_ipa_cmd(ipa_server, 'ipa group-add posixusers --desc "A POSIX group is required to log in with the user"')
-      run_ipa_cmd(ipa_server, 'ipa group-add-member posixusers --users=testuser')
+      out = run_ipa_cmd(ipa_server, admin_password, 'ipa user-add testuser --first=Test --last=User --displayname="Test User" --random')
+      run_ipa_cmd(ipa_server, admin_password, 'ipa group-add posixusers --desc "A POSIX group is required to log in with the user"')
+      run_ipa_cmd(ipa_server, admin_password, 'ipa group-add-member posixusers --users=testuser')
 
       $pass = out.stdout.split("\n").grep(/Random password/).first.split(': ').last
       puts $pass
